Remove the use of syscall within client.
The syscall package is not available on all platforms, and banned on App Engine. The use of syscall was just to
print out a syscall.Errno more clearly. However, syscall.Errno's own Error() method prints out details about the
syscall. So other than a minor change to the logging output, this changes no functionality.